feat: support markdown and HTML for description
Using `unified`, `remark`, and `rehype` to process the markdown and HTML to a sanitized and "safe" HTML
The following plugins are used:
- strip-markdown (remove/replace unwanted markdown/html element)
- remark-parse (parse MD/HTML)
- remark-breaks (add extra line breaks)
- remark-rehype (convert to MD to HTML)
- rehype-raw (parse HTML)
- rehype-sanitize (sanitize HTML)
- rehype-stringify (convert everything to HTML string)

It sill not 100% safe from XSS or other injection method but at least it is safe for now.

I guess, another way to keep it safer is to make an option where the user can enable HTML rendering or not.

Some HTML element are disabled, and some are replaced with other element.
See `filterHTMLTag` in `utils/render-markdown.ts` for more detail on what it will do.

chore: linting fix